WIP: render all the frontmatter with the YAML library
This changes how to we render frontmatter to a string, and puts
the entire front matter into a single variable of text that is
created with the YAML library.

Right now we've hard coded the frontMatterItems that are added to
the front matter, but those could be specified as an array in the
settings.